refactor: text 대신 password 매개변수 사용하도록 변경
